Make DirectSoundRenderThread::GetPlayPositionMs return double
Also make all internal storage/arithmetic related to it use doubles. The main motivation for this is that it allows us to work around some of the issues identified in #59, as it avoids 64-bit integer arithmetic altogether. doubles have 52 significand bits, so by keeping track of bytes read by DirectSound, it should be accurate up to 7 million hours or something (as opposed to about 6 hours with 32 bit integers, and even that is a probably a bit overkill actually, except for perhaps a musicdisk and perhaps some extra bits for headroom).

+++ WaveSabrePlayerLib/include/WaveSabrePlayerLib/DirectSoundRenderThread.h
@@ -17,7 +17,7 @@ namespace WaveSabrePlayerLib
 		DirectSoundRenderThread(RenderCallback callback, void *callbackData, int sampleRate, int bufferSizeMs = 1000);
 		~DirectSoundRenderThread();
 
-		int GetPlayPositionMs();
+		double GetPlayPositionMs();
 
 	private:
 		static DWORD WINAPI threadProc(LPVOID lpParameter);
@@ -34,7 +34,7 @@ namespace WaveSabrePlayerLib
 		bool shutdown;
 
 		int oldPlayCursorPos;
-		long long bytesRendered;
+		double bytesRendered;
 		LPDIRECTSOUNDBUFFER buffer;
 	};
 }
