+++ WaveSabrePlayerLib/src/PreRenderPlayer.cpp
@@ -78,7 +78,7 @@ namespace WaveSabrePlayerLib
 		if (!renderThread)
 			return 0.0;
 
-		return max(((double)renderThread->GetPlayPositionMs() - (double)playbackBufferSizeMs) / 1000.0, 0.0);
+		return max((renderThread->GetPlayPositionMs() - (double)playbackBufferSizeMs) / 1000.0, 0.0);
 	}
 
 	void PreRenderPlayer::renderCallback(SongRenderer::Sample *buffer, int numSamples, void *data)
