+++ WaveSabrePlayerLib/src/RealtimePlayer.cpp
@@ -50,7 +50,7 @@ namespace WaveSabrePlayerLib
 		if (!renderThread)
 			return 0.0;
 
-		return max(((double)renderThread->GetPlayPositionMs() - (double)bufferSizeMs) / 1000.0, 0.0);
+		return max((renderThread->GetPlayPositionMs() - (double)bufferSizeMs) / 1000.0, 0.0);
 	}
 
 	void RealtimePlayer::renderCallback(SongRenderer::Sample *buffer, int numSamples, void *data)
